Add a feature to trigger lazily loaded content
This is enabled by default, possibly it's only needed when fullPage is
true.

It will mouse wheel scroll down the page in chunks of viewportHeight,
after each scroll there's a delay of 500ms.

Example scrollY positions for a documentHeight of 9389.

0
1080
2160
4320
8640

After the final scroll we waitForNetworkIdle up to 30ms, then capture.